alice/uniproxy/server/requirements.txt
google-cloud-speech==0.36.0
pytest==5.2.0
setuptools
tornado==4.5.3
objgraph==3.4.0
protobuf>=3.6.0
flake8==3.6.0
raven
requests
ticket_parser2-py3
clickhouse_cityhash
unimemcached==0.1.28
yandex_voice_audio_decoder
PyNaCl==1.2.1
paramiko==2.4.1
ydb==0.0.37
frozendict
rtlog==0.0.22
ujson==1.35
lxml==3.5.0
